fix(shifu): satisfy Conan cache source budgets
