+++ framework/maintainability/abstraction-integrity.manifest.json
@@ -245,6 +245,15 @@
       "paths": [
         "scripts/durable-history-qualification.mjs"
       ]
+    },
+    {
+      "owner": "shifu/conan-cache",
+      "paths": [
+        "scripts/shifu-conan-hit-evidence.mjs",
+        "scripts/shifu-conan-hit-evidence.test.mjs",
+        "scripts/shifu-conan-legacy.mjs",
+        "scripts/shifu-conan-legacy.test.mjs"
+      ]
     }
   ],
   "exceptions": []
